backend/requirements.txt
flask
flask-cors
pandas
requests
scikit-learn
numpy
matplotlib
twilio
transformers
torch
python-dotenv
